starting a refactor
